test: - Provide fake prepareSimpleSearch() for better testing of Suggestor
