+++ tests/Obsidian/ObsidianSearchFacilities.test.ts
@@ -1,64 +1,6 @@
-interface SearchResult {
-    score: number;
-    matches: number[][];
-}
+import { caseInsensitiveSubstringSearch, prepareSimpleSearch } from '../__mocks__/obsidian';
 
-/**
- * A fake implementation of the function returned by prepareSimpleSearch(),
- * so we can write tests of code that calls that function.
- *
- * See https://docs.obsidian.md/Reference/TypeScript+API/prepareSimpleSearch
- * @param searchTerm
- * @param phrase
- */
-function caseInsensitiveSubstringSearch(searchTerm: string, phrase: string): SearchResult | null {
-    // Don't try and search for empty strings or just spaces:
-    if (!searchTerm.trim()) {
-        return null;
-    }
-
-    // Support multi-word search terms:
-    const searchTerms = searchTerm.split(/\s+/);
-    let matches: number[][] = [];
-
-    for (const term of searchTerms) {
-        const regex = new RegExp(term, 'gi');
-        let match;
-        let termFound = false;
-        while ((match = regex.exec(phrase)) !== null) {
-            matches.push([match.index, match.index + match[0].length]);
-            termFound = true;
-        }
-
-        // We require all search terms to be found.
-        if (!termFound) {
-            return null;
-        }
-    }
-
-    // Sort matches by start index and then by end index
-    matches = matches.sort((a, b) => {
-        if (a[0] === b[0]) {
-            return a[1] - b[1];
-        }
-        return a[0] - b[0];
-    });
-
-    return matches.length > 0
-        ? {
-              score: 0,
-              matches: matches,
-          }
-        : null;
-}
-
-export function prepareSimpleSearch(query: string): (text: string) => SearchResult | null {
-    return function (text: string): SearchResult | null {
-        return caseInsensitiveSubstringSearch(query, text);
-    };
-}
-
-describe('prepareSimpleSearch() mock', () => {
+describe('prepareSimpleSearch() fake', () => {
     it('should provide prepareSimpleSearch() function to do the search', () => {
         const searchTerm = 'hello';
         const phrase = 'hello world';
